docs: move stories for rich text editor to new ToolbarDropdown story
