+++ packages/sage-react/lib/Dropdown/Dropdown.story.jsx
@@ -2,9 +2,7 @@ import React from 'react';
 import { selectArgs } from '../story-support/helpers';
 import { SageTokens } from '../configs';
 import { Dropdown } from './Dropdown';
-import { ToolbarEditorDropdown } from './ToolbarEditorDropdown';
-import { ToolbarDropdown } from './ToolbarDropdown';
-import { defaultOptionsItems, sampleMenuItems } from './stories/story-helper';
+import { sampleMenuItems } from './stories/story-helper';
 import { CustomItemsStoryTemplate } from './stories/CustomItemsStory';
 import { CustomPanelStoryTemplate } from './stories/CustomPanelStory';
 import { BulkActionsStoryTemplate } from './stories/BulkActionsStory';
@@ -89,27 +87,6 @@ MenuWithArrow.decorators = [
     <div style={{ minHeight: 450 }}>
       { Story() }
     </div>
-  )
-];
-
-export const RichTextEditorIconOnly = (args) => (
-  <ToolbarEditorDropdown
-    triggerClassnames="sage-btn--rich-text"
-    triggerButtonSubtle={false}
-    options={defaultOptionsItems}
-    isPinned={args.isPinned}
-  />
-);
-RichTextEditorIconOnly.args = {
-  isPinned: false
-};
-RichTextEditorIconOnly.decorators = [
-  (Story) => (
-    <>
-      <div style={{ minHeight: 300 }}>
-        { Story() }
-      </div>
-    </>
   )
 ];
